+++ apps/web/src/queries/index.ts
@@ -1,0 +1,3 @@
+export { default as useCategoriesQuery } from './useCategoriesQuery';
+export { default as usePopularDrinksQuery } from './usePopularDrinksQuery';
+export { default as useRecentReviewsQuery } from './useRecentReviewsQuery';