+++ apps/web/CHANGELOG.md
@@ -2,14 +2,23 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
-### [0.0.3](https://github.com/katherinemwong/cup-trail/compare/v0.0.2...v0.0.3) (2025-08-08)
+## [0.1.0](https://github.com/Cup-Trail/cup-trail/compare/v0.0.4...v0.1.0) (2025-08-21)
 
 ### Features
 
-- add github pr comment ([7cd233c](https://github.com/katherinemwong/cup-trail/commit/7cd233c9ac05b281a292e8c7a944cacfa0ef384a))
+- Create fetch wrapper
+- Update style and layouts
+- Refactor code to use react query and react hook form
+- Refactor components and create sub-components
+
+### [0.0.3](https://github.com/Cup-Trail/cup-trail/compare/v0.0.2...v0.0.3) (2025-08-08)
+
+### Features
+
+- add github pr comment ([7cd233c](https://github.com/Cup-Trail/cup-trail/commit/7cd233c9ac05b281a292e8c7a944cacfa0ef384a))
 
 ### 0.0.2 (2025-08-08)
 
 ### Features
 
-- add eslint and typescript ([6bdd79c](https://github.com/katherinemwong/cup-trail/commit/6bdd79cf266435757d412a3e4b46da8b0f8f728e))
+- add eslint and typescript ([6bdd79c](https://github.com/Cup-Trail/cup-trail/commit/6bdd79cf266435757d412a3e4b46da8b0f8f728e))